+++ src/userconfig.hpp
@@ -151,6 +151,8 @@ public:
     //! If true, PE reads overlapping at least 'min_alignment_length' are
     //! collapsed to generate a higher quality consensus sequence.
     bool collapse;
+    //! Deterministic collapse; set equal-quality conflicting bases to N.
+    bool deterministic;
     // Allow for slipping basepairs by allowing missing bases in adapter
     unsigned shift;